requirements.txt
fastapi
uvicorn
pydantic
anthropic
langchain
langchain-anthropic
python-dotenv
numpy>=2.0
